Updates the input circuit to not take in state privately
